Remove superfluous header.
I'm not exactly certain, but I think the header was added as per advice from an automatic header trimmer (iwyu?). Either
way it does not seem to be needed and blocked compilation with clang (v9.0.0).

A note was also added about using the 'march=native' flag to automatically use atomic intrinsics. This is required for
compilation with clang (v9.0.0) within recent termux on an armv7l device.