Only link libstdc++fs if it is present.
This is a workaround for GCCv8. Previously, I was relying on it being present. It might not be though, and some
experimental Nix cross-compilation is blocked by it not being present. This commit only attempts to link when
libstdc++fs.{a,so} can be found in a system library path.